refactor(tiny-router): small refactors to match current APIs
