+++ packages/frontity/src/index.ts
@@ -1,8 +1,10 @@
-import { ConnectFunction, CreateStore } from "@frontity/types/src/connect";
 export { css, Global, keyframes } from "@emotion/core";
 export { default as styled } from "@emotion/styled";
 export { Helmet as Head } from "react-helmet";
 export { default as loadable } from "@loadable/component";
+export { observe } from "@frontity/connect";
+
+import { ConnectFunction, CreateStore } from "@frontity/types/src/connect";
 import originalConnect, {
   createStore as originalCreateStore
 } from "@frontity/connect";
