+++ packages/tiny-router/type.ts
@@ -1,22 +1,11 @@
 import Router from "@frontity/router";
-import { Derived, Action } from "@frontity/types";
-
-export type PathOrObj = string | { path: string; page: number };
+import { Action } from "frontity/types";
 
 interface TinyRouter extends Router {
   name: "@frontity/tiny-router";
-  state: {
-    frontity?: {
-      url: string;
-    };
-    router: Router["state"]["router"] & {
-      url: Derived<TinyRouter, string>;
-    };
-  };
-  actions: {
+  actions: Router["actions"] & {
     router: {
       init: Action<TinyRouter>;
-      set: Action<TinyRouter, PathOrObj>;
     };
   };
 }
